Make NumberLiteralBase available for custom number types parsers
