+++ src/Parlot/Fluent/NumberLiteral.cs
@@ -1,5 +1,4 @@
-﻿#if true
-#if !NET8_0_OR_GREATER
+﻿#if NET8_0_OR_GREATER
 using Parlot.Compilation;
 using System;
 using System.Globalization;
@@ -9,91 +8,11 @@ using System.Reflection;
 
 namespace Parlot.Fluent
 {
-    public static class NumberLiteral
+    public sealed class NumberLiteral<T> : Parser<T>, ICompilable
+        where T : INumber<T>
     {
-        public const char DefaultDecimalSeparator = '.';
-        public const char DefaultGroupSeparator = ',';
-
-        public static NumberLiteral<T> CreateNumberLiteralParser<T>(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = DefaultDecimalSeparator, char groupSeparator = DefaultGroupSeparator)
-        {
-            if (typeof(T) == typeof(byte))
-            {
-                var literal = new ByteNumberLiteral(numberOptions, decimalSeparator, groupSeparator);
-                return literal as NumberLiteral<T>;
-            }
-            else if (typeof(T) == typeof(sbyte))
-            {
-                var literal = new SByteNumberLiteral(numberOptions, decimalSeparator, groupSeparator);
-                return literal as NumberLiteral<T>;
-            }
-            else if (typeof(T) == typeof(int))
-            {
-                var literal = new IntNumberLiteral(numberOptions, decimalSeparator, groupSeparator);
-                return literal as NumberLiteral<T>;
-            }
-            else if (typeof(T) == typeof(uint))
-            {
-                var literal = new UIntNumberLiteral(numberOptions, decimalSeparator, groupSeparator);
-                return literal as NumberLiteral<T>;
-            }
-            else if (typeof(T) == typeof(long))
-            {
-                var literal = new LongNumberLiteral(numberOptions, decimalSeparator, groupSeparator);
-                return literal as NumberLiteral<T>;
-            }
-            else if (typeof(T) == typeof(ulong))
-            {
-                var literal = new ULongNumberLiteral(numberOptions, decimalSeparator, groupSeparator);
-                return literal as NumberLiteral<T>;
-            }
-            else if (typeof(T) == typeof(short))
-            {
-                var literal = new ShortNumberLiteral(numberOptions, decimalSeparator, groupSeparator);
-                return literal as NumberLiteral<T>;
-            }
-            else if (typeof(T) == typeof(ushort))
-            {
-                var literal = new UShortNumberLiteral(numberOptions, decimalSeparator, groupSeparator);
-                return literal as NumberLiteral<T>;
-            }
-            else if (typeof(T) == typeof(decimal))
-            {
-                var literal = new DecimalNumberLiteral(numberOptions, decimalSeparator, groupSeparator);
-                return literal as NumberLiteral<T>;
-            }
-            else if (typeof(T) == typeof(double))
-            {
-                var literal = new DoubleNumberLiteral(numberOptions, decimalSeparator, groupSeparator);
-                return literal as NumberLiteral<T>;
-            }
-            else if (typeof(T) == typeof(float))
-            {
-                var literal = new FloatNumberLiteral(numberOptions, decimalSeparator, groupSeparator);
-                return literal as NumberLiteral<T>;
-            }
-#if NET6_0_OR_GREATER
-            else if (typeof(T) == typeof(Half))
-            {   
-                var literal = new HalfNumberLiteral(numberOptions, decimalSeparator, groupSeparator);
-                return literal as NumberLiteral<T>;
-            }
-#endif
-            else if (typeof(T) == typeof(BigInteger))
-            {
-                var literal = new BigIntegerNumberLiteral(numberOptions, decimalSeparator, groupSeparator);
-                return literal as NumberLiteral<T>;
-            }
-            else
-            {
-                throw new NotSupportedException($"The type '{typeof(T)}' is not supported as a type argument for '{nameof(NumberLiteral<T>)}'. Only numeric types are allowed.");
-            }
-        }
-    }
-
-    public abstract class NumberLiteral<T> : Parser<T>, ICompilable
-    {
-        private static readonly MethodInfo _tryParseMethodInfo = typeof(T).GetMethod("TryParse", [typeof(string), typeof(NumberStyles), typeof(IFormatProvider), typeof(T).MakeByRefType()]);
-        private static readonly MethodInfo _rosToString = typeof(ReadOnlySpan<char>).GetMethod(nameof(ToString), []);
+        private const char DefaultDecimalSeparator = '.';
+        private const char DefaultGroupSeparator = ',';
 
         private readonly NumberOptions _numberOptions;
         private readonly char _decimalSeparator;
@@ -101,19 +20,17 @@ namespace Parlot.Fluent
         private readonly NumberStyles _numberStyles;
         private readonly CultureInfo _culture = CultureInfo.InvariantCulture;
 
-        private delegate (bool, T) TryParseDelegate();
+        private static readonly MethodInfo _tryParseMethodInfo = typeof(T).GetMethod(nameof(INumber<T>.TryParse), [typeof(ReadOnlySpan<char>), typeof(NumberStyles), typeof(IFormatProvider), typeof(T).MakeByRefType()]);
 
-        public abstract bool TryParseNumber(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider provider, out T value);
-        
-        public NumberLiteral(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = NumberLiteral.DefaultDecimalSeparator, char groupSeparator = NumberLiteral.DefaultGroupSeparator)
+        public NumberLiteral(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = DefaultDecimalSeparator, char groupSeparator = DefaultGroupSeparator)
         {
             _numberOptions = numberOptions;
             _decimalSeparator = decimalSeparator;
             _groupSeparator = groupSeparator;
             _numberStyles = _numberOptions.ToNumberStyles();
-
-            if (decimalSeparator != NumberLiteral.DefaultDecimalSeparator ||
-                groupSeparator != NumberLiteral.DefaultGroupSeparator)
+            
+            if (decimalSeparator != NumberLiterals.DefaultDecimalSeparator ||
+                groupSeparator != NumberLiterals.DefaultGroupSeparator)
             {
                 _culture = (CultureInfo)CultureInfo.InvariantCulture.Clone();
                 _culture.NumberFormat.NumberDecimalSeparator = decimalSeparator.ToString();
@@ -132,9 +49,7 @@ namespace Parlot.Fluent
             {
                 var end = context.Scanner.Cursor.Offset;
 
-                var sourceToParse = number.ToString();
-
-                if (TryParseNumber(number, _numberStyles, _culture, out T value))
+                if (T.TryParse(number, _numberStyles, _culture, out var value))
                 {
                     result.Set(start, end, value);
                     return true;
@@ -151,7 +66,7 @@ namespace Parlot.Fluent
             var result = new CompilationResult();
 
             var success = context.DeclareSuccessVariable(result, false);
-            var value = context.DeclareValueVariable(result, Expression.Default(typeof(T)));
+            var value = context.DeclareValueVariable<T>(result);
 
             // var start = context.Scanner.Cursor.Offset;
             // var reset = context.Scanner.Cursor.Position;
@@ -167,7 +82,7 @@ namespace Parlot.Fluent
             // if (context.Scanner.ReadDecimal(_numberOptions, out var numberSpan, _decimalSeparator, _groupSeparator))
             // {
             //    var end = context.Scanner.Cursor.Offset;
-            //    success = T.TryParse(numberSpan.ToString(), numberStyles, culture, out var value));
+            //    success = T.TryParse(numberSpan, numberStyles, culture, out var value));
             // }
             //
             // if (!success)
@@ -175,6 +90,9 @@ namespace Parlot.Fluent
             //    context.Scanner.Cursor.ResetPosition(begin);
             // }
             //
+
+            //var sourceToParse = Expression.Variable(typeof(ReadOnlySpan<char>), $"sourceToParse{context.NextNumber}");
+            //var sliceExpression = Expression.Assign(sourceToParse, Expression.Call(typeof(MemoryExtensions).GetMethod("AsSpan", new[] { typeof(string), typeof(int), typeof(int) }), context.Buffer(), start, Expression.Subtract(end, start)));
 
             var block =
                 Expression.IfThen(
@@ -184,7 +102,7 @@ namespace Parlot.Fluent
                         Expression.Assign(success,
                             Expression.Call(
                                 _tryParseMethodInfo,
-                                Expression.Call(numberSpan, _rosToString),
+                                numberSpan,
                                 numberStyles,
                                 culture,
                                 value)
@@ -204,190 +122,5 @@ namespace Parlot.Fluent
             return result;
         }
     }
-
-    internal sealed class ByteNumberLiteral : NumberLiteral<byte>
-    {
-        public ByteNumberLiteral(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = NumberLiteral.DefaultDecimalSeparator, char groupSeparator = NumberLiteral.DefaultGroupSeparator)
-            : base(numberOptions, decimalSeparator, groupSeparator)
-        {
-
-        }
-
-        public override bool TryParseNumber(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider provider, out byte value)
-        {
-            return byte.TryParse(s.ToString(), style, provider, out value );
-        }
-    }
-
-    internal sealed class SByteNumberLiteral : NumberLiteral<sbyte>
-    {
-        public SByteNumberLiteral(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = NumberLiteral.DefaultDecimalSeparator, char groupSeparator = NumberLiteral.DefaultGroupSeparator)
-            : base(numberOptions, decimalSeparator, groupSeparator)
-        {
-
-        }
-
-        public override bool TryParseNumber(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider provider, out sbyte value)
-        {
-            return sbyte.TryParse(s.ToString(), style, provider, out value);
-        }
-    }
-
-    internal sealed class IntNumberLiteral : NumberLiteral<int>
-    {
-        public IntNumberLiteral(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = NumberLiteral.DefaultDecimalSeparator, char groupSeparator = NumberLiteral.DefaultGroupSeparator)
-            : base(numberOptions, decimalSeparator, groupSeparator)
-        {
-
-        }
-
-        public override bool TryParseNumber(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider provider, out int value)
-        {
-            return int.TryParse(s.ToString(), style, provider, out value);
-        }
-    }
-
-    internal sealed class UIntNumberLiteral : NumberLiteral<uint>
-    {
-        public UIntNumberLiteral(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = NumberLiteral.DefaultDecimalSeparator, char groupSeparator = NumberLiteral.DefaultGroupSeparator)
-            : base(numberOptions, decimalSeparator, groupSeparator)
-        {
-
-        }
-
-        public override bool TryParseNumber(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider provider, out uint value)
-        {
-            return uint.TryParse(s.ToString(), style, provider, out value);
-        }
-    }
-
-    internal sealed class LongNumberLiteral : NumberLiteral<long>
-    {
-        public LongNumberLiteral(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = NumberLiteral.DefaultDecimalSeparator, char groupSeparator = NumberLiteral.DefaultGroupSeparator)
-            : base(numberOptions, decimalSeparator, groupSeparator)
-        {
-
-        }
-
-        public override bool TryParseNumber(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider provider, out long value)
-        {
-            return long.TryParse(s.ToString(), style, provider, out value);
-        }
-    }
-
-    internal sealed class ULongNumberLiteral : NumberLiteral<ulong>
-    {
-        public ULongNumberLiteral(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = NumberLiteral.DefaultDecimalSeparator, char groupSeparator = NumberLiteral.DefaultGroupSeparator)
-            : base(numberOptions, decimalSeparator, groupSeparator)
-        {
-
-        }
-
-        public override bool TryParseNumber(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider provider, out ulong value)
-        {
-            return ulong.TryParse(s.ToString(), style, provider, out value);
-        }
-    }
-
-    internal sealed class ShortNumberLiteral : NumberLiteral<short>
-    {
-        public ShortNumberLiteral(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = NumberLiteral.DefaultDecimalSeparator, char groupSeparator = NumberLiteral.DefaultGroupSeparator)
-            : base(numberOptions, decimalSeparator, groupSeparator)
-        {
-
-        }
-
-        public override bool TryParseNumber(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider provider, out short value)
-        {
-            return short.TryParse(s.ToString(), style, provider, out value);
-        }
-    }
-
-    internal sealed class UShortNumberLiteral : NumberLiteral<ushort>
-    {
-        public UShortNumberLiteral(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = NumberLiteral.DefaultDecimalSeparator, char groupSeparator = NumberLiteral.DefaultGroupSeparator)
-            : base(numberOptions, decimalSeparator, groupSeparator)
-        {
-
-        }
-
-        public override bool TryParseNumber(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider provider, out ushort value)
-        {
-            return ushort.TryParse(s.ToString(), style, provider, out value);
-        }
-    }
-
-    internal sealed class DecimalNumberLiteral : NumberLiteral<decimal>
-    {
-        public DecimalNumberLiteral(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = NumberLiteral.DefaultDecimalSeparator, char groupSeparator = NumberLiteral.DefaultGroupSeparator)
-            : base(numberOptions, decimalSeparator, groupSeparator)
-        {
-
-        }
-
-        public override bool TryParseNumber(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider provider, out decimal value)
-        {
-            return decimal.TryParse(s.ToString(), style, provider, out value);
-        }
-    }
-
-    internal sealed class DoubleNumberLiteral : NumberLiteral<double>
-    {
-        public DoubleNumberLiteral(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = NumberLiteral.DefaultDecimalSeparator, char groupSeparator = NumberLiteral.DefaultGroupSeparator)
-            : base(numberOptions, decimalSeparator, groupSeparator)
-        {
-
-        }
-
-        public override bool TryParseNumber(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider provider, out double value)
-        {
-            return double.TryParse(s.ToString(), style, provider, out value);
-        }
-    }
-
-    internal sealed class FloatNumberLiteral : NumberLiteral<float>
-    {
-        public FloatNumberLiteral(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = NumberLiteral.DefaultDecimalSeparator, char groupSeparator = NumberLiteral.DefaultGroupSeparator)
-            : base(numberOptions, decimalSeparator, groupSeparator)
-        {
-
-        }
-
-        public override bool TryParseNumber(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider provider, out float value)
-        {
-            return float.TryParse(s.ToString(), style, provider, out value);
-        }
-    }
-
-#if NET6_0_OR_GREATER
-    internal sealed class HalfNumberLiteral : NumberLiteral<Half>
-    {
-        public HalfNumberLiteral(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = NumberLiteral.DefaultDecimalSeparator, char groupSeparator = NumberLiteral.DefaultGroupSeparator)
-            : base(numberOptions, decimalSeparator, groupSeparator)
-        {
-
-        }
-
-        public override bool TryParseNumber(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider provider, out Half value)
-        {
-            return Half.TryParse(s.ToString(), style, provider, out value);
-        }
-    }
-#endif
-
-    internal sealed class BigIntegerNumberLiteral : NumberLiteral<BigInteger>
-    {
-        public BigIntegerNumberLiteral(NumberOptions numberOptions = NumberOptions.Number, char decimalSeparator = NumberLiteral.DefaultDecimalSeparator, char groupSeparator = NumberLiteral.DefaultGroupSeparator)
-            : base(numberOptions, decimalSeparator, groupSeparator)
-        {
-
-        }
-
-        public override bool TryParseNumber(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider provider, out BigInteger value)
-        {
-            return BigInteger.TryParse(s.ToString(), style, provider, out value);
-        }
-    }
 }
-#endif
 #endif
